ids and block numbers are bigints too now fixed a UMR in psql_read_buf

+++ pgsql.h
@@ -72,28 +72,28 @@ int psql_rollback( PGconn *conn );
 
 int psql_path_to_id( PGconn *conn, const char *path );
 
-int psql_read_meta( PGconn *conn, const int id, const char *path, PgMeta *meta );
+int psql_read_meta( PGconn *conn, const int64_t id, const char *path, PgMeta *meta );
 
 int psql_read_meta_from_path( PGconn *conn, const char *path, PgMeta *meta );
 
-int psql_write_meta( PGconn *conn, const int id, const char *path, PgMeta meta );
+int psql_write_meta( PGconn *conn, const int64_t id, const char *path, PgMeta meta );
 
-int psql_create_file( PGconn *conn, const int parent_id, const char *path, const char *new_file, PgMeta meta );
+int psql_create_file( PGconn *conn, const int64_t parent_id, const char *path, const char *new_file, PgMeta meta );
 
-int psql_read_buf( PGconn *conn, const int id, const char *path, char *buf, const off_t offset, const size_t len, int verbose );
+int psql_read_buf( PGconn *conn, const int64_t id, const char *path, char *buf, const off_t offset, const size_t len, int verbose );
 
-int psql_readdir( PGconn *conn, const int parent_id, void *buf, fuse_fill_dir_t filler );
+int psql_readdir( PGconn *conn, const int64_t parent_id, void *buf, fuse_fill_dir_t filler );
 
-int psql_create_dir( PGconn *conn, const int parent_id, const char *path, const char *new_dir, PgMeta meta );
+int psql_create_dir( PGconn *conn, const int64_t parent_id, const char *path, const char *new_dir, PgMeta meta );
 
-int psql_delete_dir( PGconn *conn, const int id, const char *path );
+int psql_delete_dir( PGconn *conn, const int64_t id, const char *path );
 
-int psql_delete_file( PGconn *conn, const int id, const char *path );
+int psql_delete_file( PGconn *conn, const int64_t id, const char *path );
 
-int psql_write_buf( PGconn *conn, const int id, const char *path, const char *buf, const off_t offset, const size_t len, int verbose );
+int psql_write_buf( PGconn *conn, const int64_t id, const char *path, const char *buf, const off_t offset, const size_t len, int verbose );
 
-int psql_truncate( PGconn *conn, const int id, const char *path, const off_t offset );
+int psql_truncate( PGconn *conn, const int64_t id, const char *path, const off_t offset );
 
-int psql_rename( PGconn *conn, const int from_id, const int from_parent_id, const int to_parent_id, const char *rename_to, const char *from, const char *to );
+int psql_rename( PGconn *conn, const int64_t from_id, const int64_t from_parent_id, const int64_t to_parent_id, const char *rename_to, const char *from, const char *to );
 
 #endif
